update zip list struc
