+++ linked_list/README.md
@@ -13,7 +13,6 @@ A linked list is a linear data structure, in which the elements are not stored a
   - insert before
   - insert after
   - kth_form_end
-  - zip lists
 - Testing
 
 ## Approach & Efficiency
@@ -28,7 +27,6 @@ Time complexity:
  - insert after : (n)
  - insert before : (n)
  - kth_form_end : (n)
- - zip_lists : (n)
 
 Space complexity:
  - insert: (n)
@@ -38,7 +36,7 @@ Space complexity:
  - insert after : (1)
  - insert before : (1)
  - kth_form_end (1)
- - zip_lists : (1)
+
 
 
 ## Whiteboard Process
@@ -50,8 +48,6 @@ Space complexity:
   * ![insert after](insert%20after.jpg)
 * kth_form_end: 
   * ![kth_form_end](Linked_kth.jpg)
-* zip_lists: 
-  * ![zip_lists](linked-list-zip.jpg)
 ## API
 <!-- Description of each method publicly available to your Linked List -->
 > Insert:
