+++ linked_list/zip_list/zip_list.py
