+++ tests/test_linked_list.py
@@ -1,6 +1,6 @@
 import pytest
-from linked_list.linked_list import Node, LinkedList
-from linked_list.zip_list import zip_lists
+from linked_list.linked_list import LinkedList
+from linked_list.zip_list.zip_list import zip_lists
 
 #  instantiate an empty linked list
 def test_linked_list(ll):
@@ -180,10 +180,6 @@ def test_kth_in_middle(ll):
     expected = "12"
     assert expected == str(actual)
 
-def test_zip_listes(ll_one, ll_two):
-    actual = zip_lists(ll_one,ll_two).to_string()
-    expected = '{ 1 } ->{ 5 } ->{ 3 } ->{ 9 } ->{ 2 } ->{ 4 } -> NULL'
-    assert expected == actual
 
 @pytest.fixture
 def ll():
@@ -211,18 +207,3 @@ def ll_size1():
     return ll_size1
 
 
-@pytest.fixture
-def ll_one():
-    ll_one = LinkedList()
-    ll_one.insert(2)
-    ll_one.insert(3)
-    ll_one.insert(1)
-    return ll_one
-
-@pytest.fixture
-def ll_two():
-    ll_two = LinkedList()
-    ll_two.insert(4)
-    ll_two.insert(9)
-    ll_two.insert(5)
-    return ll_two
